+++ emg3d/__init__.py
@@ -25,9 +25,10 @@ sped-up through jitted ``numba``-functions.
 
 from emg3d import utils
 from emg3d import solver
+from emg3d.solver import solve
 from emg3d.utils import Report
 
-__all__ = ['solver', 'utils', 'Report']
+__all__ = ['solve', 'solver', 'utils', 'Report']
 
 # Version is sorted out in utils, so we can easier use it within the package
 # itself.