+++ tests/test_njitted.py
@@ -33,7 +33,7 @@ def test_amat_x():
     vmodel = utils.VolumeModel(grid, model, sfield)
 
     # Run two iterations to get a e-field
-    efield = solver.solver(grid, model, sfield, maxit=2, verb=1)
+    efield = solver.solve(grid, model, sfield, maxit=2, verb=1)
 
     # amat_x
     rr1 = utils.Field(grid)
@@ -114,7 +114,7 @@ def test_gauss_seidel():
         vmodel = utils.VolumeModel(grid, model, sfield)
 
         # Run two iterations to get some e-field.
-        efield = solver.solver(grid, model, sfield, maxit=2, verb=1)
+        efield = solver.solve(grid, model, sfield, maxit=2, verb=1)
 
         inp = (sfield.fx, sfield.fy, sfield.fz, vmodel.eta_x, vmodel.eta_y,
                vmodel.eta_z, vmodel.zeta, grid.hx, grid.hy, grid.hz, nu)